requirements.txt
streamlit>=1.53.0
pandas>=2.3.3
numpy>=2.4.1
plotly>=6.5.2
google-generativeai>=0.4.0
sqlalchemy>=2.0.45
psycopg2-binary>=2.9.11
bcrypt>=4.0.1
openpyxl>=3.1.5
matplotlib>=3.10.8
seaborn>=0.13.2
scikit-learn>=1.8.0
scipy>=1.17.0
